BRANCH 4.1: - more updates for Mongodb dealing with search_type, search_code/id relationship. - updated version to b02

+++ src/pyasm/web/cherrypy_startup.py
@@ -158,17 +158,18 @@ class CherryPyStartup(object):
 
         try:
             sql = DbContainer.get("sthpw")
-            # before batch, clean up the ticket with a NULL code
-            if os.getenv('TACTIC_MODE') != 'production':
-                sql.do_update('DELETE from "ticket" where "code" is NULL;')
-            else:
-                start_port = Config.get_value("services", "start_port")
-                if start_port:
-                    start_port = int(start_port)
+            if sql.get_database_type() != "MongoDb":
+                # before batch, clean up the ticket with a NULL code
+                if os.getenv('TACTIC_MODE') != 'production':
+                    sql.do_update('DELETE from "ticket" where "code" is NULL;')
                 else:
-                    start_port = 8081
-                if port and int(port) == start_port:
-                     sql.do_update('DELETE from "ticket" where "code" is NULL;')
+                    start_port = Config.get_value("services", "start_port")
+                    if start_port:
+                        start_port = int(start_port)
+                    else:
+                        start_port = 8081
+                    if port and int(port) == start_port:
+                         sql.do_update('DELETE from "ticket" where "code" is NULL;')
         except DatabaseException, e:
             # TODO: need to work on this
             print "ERROR: could not connect to [sthpw] database"
